+++ NEWS.md
@@ -1,5 +1,11 @@
 # tensorflow (development version)
 
+- The package now Suggest 'keras3' instead of 'keras'
+- `install_tensorflow()` installs TensorFlow v2.16 by default.
+- If `install_tensorflow()` detects a GPU on Linux, it will automatically 
+  install the cuda package and configure required symlinks for cudnn.
+
+
 # tensorflow 2.15.0
 
 - `install_tensorflow()` installs TensorFlow v2.15 by default